Added TypeScript support for jest
