+++ package.json
@@ -18,8 +18,10 @@
     "yargs": "^17.5.1"
   },
   "devDependencies": {
-    "jest": "^28.1.0",
+    "@types/jest": "^28.1.1",
+    "jest": "^28.1.1",
     "nodemon": "^2.0.16",
+    "ts-jest": "^28.0.4",
     "ts-node": "^10.8.1",
     "typescript": "^4.7.3"
   }
