--- pokedex.test.ts
@@ -1,28 +1,0 @@
-const { PokedexModule } = require('./src/pokedex-module');
-
-const tests = [
-  {
-    name: 'get charizard - english',
-    language: "en",
-    input: "charizard",
-    expectedOutput: "Charizard"
-  },
-  {
-    name: 'get charizard - german',
-    language: "de",
-    input: "charizard",
-    expectedOutput: "Glurak"
-  }, 
-];
-
-tests.forEach(t => {
-  test(t.name, async () => {
-    let lang = t.language;
-
-    const data = await Pokedex.getPokemonSpecies(t.input);
-
-    let name = Pokedex.getPokemonName(data, lang)
-    
-    expect(name).toBe(t.expectedOutput);
-    });
-});
